feat: created Populator to Glue BusinessComponents with GH
