+++ AdSecGHTests/Helpers/DataConvertorTest.cs
@@ -1,61 +1,30 @@
-﻿using System;
+﻿using Oasys.GH.Helpers;
 
-using Oasys.Business;
-
-using Attribute = Oasys.Business.Attribute;
+using Xunit;
 
 namespace AdSecGHTests.Helpers {
-  // [Collection("GrasshopperFixture collection")]
-  // public class DataConvertorTest {
-  //   private readonly DummyComponent _component;
-  //
-  //   public DataConvertorTest() {
-  //     _component = new DummyComponent();
-  //   }
-  //
-  //   [Fact]
-  //   public void ShouldHaveTheSameNumberOfInputs() {
-  //     Assert.Single(_component.Params.Input);
-  //   }
-  //
-  //   [Fact]
-  //   public void ShouldPassTheName() {
-  //     Assert.Equal("Alpha", _component.Params.Input[0].Name);
-  //   }
-  //
-  //   [Fact]
-  //   public void ShouldPassTheNickname() {
-  //     Assert.Equal("A", _component.Params.Input[0].NickName);
-  //   }
-  // }
+  [Collection("GrasshopperFixture collection")]
+  public class DataConvertorTest {
+    private readonly DummyComponent _component;
 
-  public class DummyBusiness : IBusinessComponent {
-
-    public DoubleParameter Alpha { get; set; } = new DoubleParameter {
-      Name = "Alpha",
-      NickName = "A",
-      Description = "Alpha description",
-    };
-    public DoubleParameter Beta { get; set; }
-
-    public Attribute[] GetAllInputAttributes() {
-      return new[] {
-        Alpha,
-      };
+    public DataConvertorTest() {
+      _component = new DummyComponent();
     }
 
-    public Attribute[] GetAllOutputAttributes() {
-      return new[] {
-        Beta,
-      };
+    [Fact]
+    public void ShouldHaveTheSameNumberOfInputs() {
+      Assert.Single(_component.Params.Input);
     }
 
-    public void UpdateInputValues(params object[] values) { throw new NotImplementedException(); }
+    [Fact]
+    public void ShouldPassTheName() {
+      Assert.Equal("Alpha", _component.Params.Input[0].Name);
+    }
 
-    public void Compute() { Beta.Value = Alpha.Value * 2; }
+    [Fact]
+    public void ShouldPassTheNickname() {
+      Assert.Equal("A", _component.Params.Input[0].NickName);
+    }
   }
 
-  public class DummyComponent : BusinessComponentGlue<DummyBusiness> {
-    public DummyComponent() : base("name", "nickname", "description", "category", "subcategory") { }
-  }
 }
